feat: Add image conversion and display functionalities
- Implemented a new script `convert_new.py` for bulk-converting various image formats (FITS, PNG, JPEG, TIFF) into a sharded Zarr store, including support for resizing and metadata handling.
- Introduced `display_sample_images.py` to visualize random samples from the Zarr store, with options to save the output.
- Removed redundant sample image display function from `inspect.py`.
- Updated `pyproject.toml` to bump version to 0.3.0 and added `scikit-image` as a dependency.
