+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "images_to_zarr"
-version = "0.2.0"
+version = "0.3.0"
 description = "Tiny Python module to bulk-convert large amounts of images into zarr files"
 readme = "README.md"
 license = { file = "LICENSE" }
@@ -26,6 +26,7 @@ dependencies = [
     "loguru",
     "numpy",
     "pandas",
+    "scikit-image",
     "tqdm",
     "zarr>=3.0.0b0",
 ]
